requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
brotlipy==0.7.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-admin-persian-fonts==0.2
django-celery-beat==2.4.0
django-hashid-field==3.3.7
django-import-export==3.1.0
django-jalali-date==1.0.1
django-reorder-admin==0.3.1
django-timezone-field==5.0
environs==9.5.0
et-xmlfile==1.1.0
hashids==1.3.1
idna==3.4
ippanel==2.0.7
isodate==0.6.1
jdatetime==4.1.0
kombu==5.2.4
lxml==4.9.2
MarkupPy==1.14
marshmallow==3.19.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
Pillow==9.4.0
platformdirs==3.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
